requirements.txt
SQLAlchemy==1.3.17
flask-sqlalchemy==2.4.3
Flask-Seeder==1.1.1
WTForms==2.3.1
flask-wtf==0.14.3
bcrypt==3.1.7
cffi==1.14.0
flask-bcrypt==0.7.1
pycparser==2.20